refactor: rename decision-instance v2 mocks
